steps towards being able to do a slick user dao
Former-commit-id: 87705b4cb0106727a4651c661ea3dfbcff73d70c [formerly 35b83387b67272e1207f764dc4d65cc4e1e28d58] [formerly 0160e5b052456d530a3acece07d3f3517a58a86f [formerly 257258b97d5a9d3116e08e6012fa824da9ba925d]]
Former-commit-id: 2e580832423528faa9e18f852f5f12ff786f2a2f [formerly 52602ba09669b43f7ea4c49c9aa2a4421ad3a4d6]
Former-commit-id: 82707691fa1d079ebe21b183b26546fecc8ebdec

+++ src/mitll/langtest/shared/exercise/AudioAttribute.java
@@ -42,7 +42,7 @@ public class AudioAttribute implements IsSerializable, UserAndTime {
   private int uniqueID;
   private String audioRef;
   private String exid;
-  private long userid;
+  private int userid;
   private long timestamp;
   private long durationInMillis;
   /**
@@ -70,7 +70,7 @@ public class AudioAttribute implements IsSerializable, UserAndTime {
    * @see mitll.langtest.server.database.AudioDAO#getAudioAttribute
    * @see mitll.langtest.server.database.AudioDAO#getResultsForQuery(java.sql.Connection, java.sql.PreparedStatement)
    */
-  public AudioAttribute(int uniqueID, long userid,
+  public AudioAttribute(int uniqueID, int userid,
                         String exid,
                         String audioRef,
                         long timestamp, long durationInMillis, String type, MiniUser user, String transcript) {
@@ -244,7 +244,7 @@ public class AudioAttribute implements IsSerializable, UserAndTime {
     this.user = user;
   }
 
-  public long getUserid() {
+  public int getUserid() {
     return userid;
   }
